requirements.txt
librosa
numpy
scikit-learn
soundfile
argparse
pydub
python-dotenv
lyricsgenius
langdetect
requests
pandas
tabulate
matplotlib
essentia